# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Fabel API
  version: 0.0.1
  description: |
    SSE endpoint POST /api/conversations/{id}/send is not modeled here.
    It is registered manually on the chi router.

paths:
  # ── Auth ──────────────────────────────────────────────
  /api/auth/register:
    post:
      operationId: register
      summary: Register a new user
      description: |
        First registered user becomes admin. Subsequent registrations
        require allow_registration=true in server settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
        "403":
          description: Registration is disabled

  /api/auth/login:
    post:
      operationId: login
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials

  /api/auth/logout:
    post:
      operationId: logout
      summary: Log out (destroy session)
      responses:
        "204":
          description: Logged out

  /api/auth/me:
    get:
      operationId: getMe
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── Admin ─────────────────────────────────────────────
  /api/admin/users:
    get:
      operationId: listUsers
      summary: List all users (admin only)
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        "403":
          description: Not admin

  /api/admin/users/{id}:
    patch:
      operationId: updateUser
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
        "403":
          description: Not admin
        "404":
          description: User not found

  /api/admin/settings:
    get:
      operationId: getSettings
      summary: Get server settings (admin only)
      responses:
        "200":
          description: Server settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerSettings"
        "403":
          description: Not admin
    patch:
      operationId: updateSettings
      summary: Update server settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerSettings"
        "403":
          description: Not admin

  # ── App ───────────────────────────────────────────────
  /api/bootstrap:
    get:
      operationId: getBootstrap
      summary: Initial data load
      responses:
        "200":
          description: Characters, presets, and optional seeded conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapResponse"

  /api/conversations:
    post:
      operationId: createConversation
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "201":
          description: Created conversation with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetConversationResponse"

  /api/conversations/{id}:
    get:
      operationId: getConversation
      summary: Get conversation with details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with character, preset, and messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetConversationResponse"
        "404":
          description: Conversation not found

  /api/conversations/{id}/prompt:
    get:
      operationId: getConversationPrompt
      summary: Get constructed prompt for a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Constructed prompt messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptResponse"
        "404":
          description: Conversation not found

components:
  schemas:
    # ── Auth schemas ──────────────────────────────────────
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 32
        password:
          type: string
          minLength: 8
        display_name:
          type: string
          maxLength: 64

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    User:
      type: object
      required: [id, username, display_name, role, created_at]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [admin, user]
        created_at:
          type: string

    UserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"

    UpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, user]

    ServerSettings:
      type: object
      properties:
        allow_registration:
          type: boolean

    # ── App schemas ───────────────────────────────────────
    Character:
      type: object
      required: [id, name, description, personality, scenario, first_mes, mes_example]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        personality:
          type: string
        scenario:
          type: string
        first_mes:
          type: string
        mes_example:
          type: string

    Preset:
      type: object
      required: [id, name, model, temperature, max_tokens, system_prompt]
      properties:
        id:
          type: integer
        name:
          type: string
        model:
          type: string
        temperature:
          type: number
          format: double
        max_tokens:
          type: integer
        system_prompt:
          type: string

    Message:
      type: object
      required: [id, conversation_id, role, content, created_at]
      properties:
        id:
          type: integer
        conversation_id:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        parent_id:
          type: integer
          nullable: true
        active_child_id:
          type: integer
          nullable: true
        created_at:
          type: string

    ConversationDetail:
      type: object
      required: [id, character_id, preset_id, created_at, character, preset, messages]
      properties:
        id:
          type: integer
        character_id:
          type: integer
        preset_id:
          type: integer
        created_at:
          type: string
        character:
          $ref: "#/components/schemas/Character"
        preset:
          $ref: "#/components/schemas/Preset"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"

    CreateConversationRequest:
      type: object
      required: [character_id, preset_id]
      properties:
        character_id:
          type: integer
        preset_id:
          type: integer

    BootstrapResponse:
      type: object
      required: [characters, presets]
      properties:
        characters:
          type: array
          items:
            $ref: "#/components/schemas/Character"
        presets:
          type: array
          items:
            $ref: "#/components/schemas/Preset"
        seeded_conversation:
          $ref: "#/components/schemas/ConversationDetail"

    GetConversationResponse:
      type: object
      required: [conversation]
      properties:
        conversation:
          $ref: "#/components/schemas/ConversationDetail"

    PromptMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string

    PromptResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/PromptMessage"